+++ code/server/Plateau.h
@@ -12,27 +12,27 @@
 
 class Plateau {
 public:
-    Plateau(size_t width, size_t height);
+    Plateau(unsigned int width, unsigned int height);
 
-    Case& getCase(size_t x, size_t y);
-    const Case& getCase(size_t x, size_t y) const;
+    Case& getCase(unsigned int x, unsigned int y);
+    const Case& getCase(unsigned int x, unsigned int y) const;
 
-    bool isWalkable(size_t x, size_t y) const;
-    bool isOccupied(size_t x, size_t y, uint32_t excludeId,
+    bool isWalkable(unsigned int x, unsigned int y) const;
+    bool isOccupied(unsigned int x, unsigned int y, uint32_t excludeId,
                     const std::vector<Player>& players) const;
-    bool isInside(size_t x, size_t y) const;
+    bool isInside(unsigned int x, unsigned int y) const;
 
 
     void print() const;
     void printWithPlayers(const std::vector<Player>& players) const;
 
-    size_t getWidth() const { return width; }
-    size_t getHeight() const { return height; }
+    unsigned int getWidth() const { return width; }
+    unsigned int getHeight() const { return height; }
     BoardData toData() const;
 
 private:
-    size_t width;
-    size_t height;
+    unsigned int width;
+    unsigned int height;
     gf::Array2D<Case> grid;
 
 };
